allow easy disable of sif creation

+++ make-container.sh
@@ -2,6 +2,8 @@
 
 set -e
 
+makesif=0
+
 if [ -z "$1" ] ; then
   echo "Error - you must provide a name for your container"
   echo "example: $0 my-container"
@@ -23,23 +25,26 @@ bin/make-install-script.sh "$PWD"
 
 docker build -t "$container_name" .
 
-savefile=$(mktemp)
-savefile_name=$(basename "$savefile")
-savefile_dir=$(dirname "$savefile")
-
-docker save "$container_name" > "$savefile"
-
-d="$PWD"
-
-cd singularity
-
-rand=$(dd if=/dev/urandom count=1 bs=512 2>/dev/null | openssl sha1 )
-singularity_tag="rbuilder-sif-singularity-${rand}" 
-docker build -t "$singularity_tag" .
-
-cd "$d"
-
-docker run -v "$savefile_dir:/in" -v "$PWD:/out" -it "$singularity_tag" bash -c "singularity build /out/${container_name}.sif docker-archive://in/${savefile_name}"
-
-set +e
-rm "$savefile" 
+if [ "$makesif" -ne 0 ] ; then
+  
+  savefile=$(mktemp)
+  savefile_name=$(basename "$savefile")
+  savefile_dir=$(dirname "$savefile")
+  
+  docker save "$container_name" > "$savefile"
+  
+  d="$PWD"
+  
+  cd singularity
+  
+  rand=$(dd if=/dev/urandom count=1 bs=512 2>/dev/null | openssl sha1 )
+  singularity_tag="rbuilder-sif-singularity-${rand}" 
+  docker build -t "$singularity_tag" .
+  
+  cd "$d"
+  
+  docker run -v "$savefile_dir:/in" -v "$PWD:/out" -it "$singularity_tag" bash -c "singularity build /out/${container_name}.sif docker-archive://in/${savefile_name}"
+  
+  set +e
+  rm "$savefile" 
+fi
